WIP: reorganize token data and make filename more generic
This is demonstrating how we can use Style Dictionary to convert our design-tokens.json file into something we can use in our custom Stylelint rules. In this example the Storybook tokens table is using the same data, but we could create a separate file/Style Dictionary format if we think that's best.

Differential Revision: https://phabricator.services.mozilla.com/D254061